+++ test/custom_sound_manager.dart
@@ -1,5 +1,5 @@
 /// Provides the [CustomSoundManager] class.
-import 'package:dart_synthizer/context.dart';
+import 'package:dart_synthizer/dart_synthizer.dart';
 import 'package:ziggurat/ziggurat.dart';
 import 'package:ziggurat_sounds/ziggurat_sounds.dart';